silence compiler warnings about comparison of signed and unsigned variables

+++ src/lib/simulator/electronics/gpsimprocessor.cpp
@@ -771,14 +771,14 @@ RegisterSet::RegisterSet( pic_processor * picProcessor )
 
 RegisterSet::~RegisterSet()
 {
-	for ( unsigned i = 0; i < m_registers.size(); ++i )
+	for (int i = 0; i < m_registers.size(); ++i )
 		delete m_registers[i];
 }
 
 
 RegisterInfo * RegisterSet::fromAddress( unsigned address )
 {
-	return (address < m_registers.size()) ? m_registers[address] : 0l;
+	return ((int)address < m_registers.size()) ? m_registers[address] : 0l;
 }
 
 
@@ -803,7 +803,7 @@ RegisterInfo * RegisterSet::fromName( const QString & name )
 
 void RegisterSet::update()
 {
-	for ( unsigned i = 0; i < m_registers.size(); ++i )
+	for (int i = 0; i < m_registers.size(); ++i )
 		m_registers[i]->update();
 }
 //END class RegisterSet
